requirements.txt
# local package
-e .

# external requirements
mysql-connector-python

